requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
boto3==1.34.34
botocore==1.34.34
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.18.1
cloup==3.0.5
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
duckduckgo_search==6.1.9
email_validator==2.2.0
executing==2.0.1
fastai==2.7.15
fastapi==0.109.2
fastapi-cli==0.0.4
fastbook==0.0.29
fastcore==1.5.48
fastdownload==0.0.7
fastjsonschema==2.20.0
fastprogress==1.0.3
filelock==3.15.4
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
glcontext==2.5.0
graphviz==0.20.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
img2vec-pytorch==1.0.1
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isoduration==20.11.0
isosurfaces==0.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab-widgets==1.1.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
manim==0.18.1
ManimPango==0.5.0
mapbox-earcut==1.0.1
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mock==5.1.0
moderngl==5.10.0
moderngl-window==2.4.6
mpmath==1.3.0
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
path==16.14.0
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycairo==1.26.1
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydub==0.25.1
pyglet==2.0.16
Pygments==2.18.0
pyparsing==3.1.2
pyreqwest_impersonate==0.4.9
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
screeninfo==0.8.1
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
skia-pathops==0.8.0.post1
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
srt==3.5.3
stack-data==0.6.3
starlette==0.36.3
svgelements==1.9.6
sympy==1.12.1
tbb==2021.13.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.1+cu118
torchinfo==1.8.0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.27.0.post1
virtualenv==20.26.3
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==3.6.6
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
